Added jumping
You can now jump and do a flip in mid-air by pressing Space. I also increased veer force (so that you can move side to side faster), gravity (to make jumping work better), and max speed. I made some changes to HandleVeering to help keep the car from going off of the edge of the road and hitting fire hydrants. There are still some issues with the car floating in mid-air, or not jumping at all, and with the sound not changing pitch, so feel free to play around with my changes and make them work better.